+++ lib/dart_class_mapper.dart
@@ -1,3 +1,6 @@
 export 'src/create_map.dart';
 export 'src/get_mapper.dart';
+export 'src/mapper/mapper.dart';
+export 'src/mapper/mapper_extensions.dart';
+export 'src/mapper/mapper_utils.dart';
 export 'src/exceptions/dart_class_mapper_exception.dart';